requirements-dev.txt
# Development dependencies
-r requirements.txt

# Code formatting
black==23.12.1
isort==5.13.2

# Linting
flake8==7.0.0
mypy==1.8.0

# Pre-commit hooks
pre-commit==3.6.0

# Testing extras
pytest-asyncio==0.23.2
pytest-mock==3.12.0
